Add files via upload
These are all of the functions and classes that are likely to be part of the final v1 program. All are incomplete, most need to have the logic in their functions re-written and will also have to be integrated with more classes that have yet to be written.

Currently finishing up instating naming conventions (mostly based on PEP8 guidelines), deleting irrelevant code, and scribbling on a legal pad about what new classes I need to write and how it will affect the responsibilities of this collection of classes I have already written.